requirements.txt
pillow
streamlit
streamlit-cropper
